requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-telegram-bot>=13.7
pymongo>=3.12.0
python-dotenv>=0.19.0
requests>=2.26.0
fpdf2>=2.5.6
gunicorn>=20.1.0